moved MapboxRouteLineApiExtensions to a stable package
