+++ libnavui-maps/src/main/java/com/mapbox/navigation/ui/maps/route/line/MapboxRouteLineApiExtensions.kt
@@ -1,4 +1,4 @@
-package com.mapbox.navigation.ui.maps.internal.route.line
+package com.mapbox.navigation.ui.maps.route.line
 
 import com.mapbox.api.directions.v5.models.DirectionsRoute
 import com.mapbox.bindgen.Expected
